Fix VMWare detection
The original code was reading `dmesg` output. That was fine immediately
after boot, but various events would eventually overwrite the dmesg
buffer and break VMWare detection.

This was masked because the code in readTagOrKnob was mistakenly
checking the existence of the inVMware function instead of its output.

I didn't realize this till I refactored and adding more spammy debug
messages and realized I wasn't seeing it declare it was in VMware on the
test instance.

And of course the old method would work when we tested it on fresh instances.