Deal with more vSphere shenanigans.
* Add some missing docstrings and clean up some existing ones
* Add loadVSphereSettings()
* Add writeVSphereSettings()
* Add detectVSphereHost()
* Update readVirtualMachineTag()

Originally, we walked the list of hypervisors every time we needed to
read a VM tag, and that incurred a delay penalty when we couldn't
connect.

Now, when we try to read any VM tag, we attempt to read the hypervisor
connection information from /etc/knobs/vsphere.toml. If it isn't
present, we call detectVSphereHost() - it determines the correct
hypervisor and writes the cache to /etc/knobs/vsphere.toml for
future tag reads.

This makes us only pay the timeout penalty the first time we run
sourdough on a machine, subsequent runs will use the cached data in the
knob file.